Moved some logic to BLL from DAL

+++ Knowledge_testing_system/Controllers/UserController.cs
@@ -1,4 +1,5 @@
-﻿using KnowledgeTestingSystemBLL.Entities;
+﻿using AutoMapper;
+using KnowledgeTestingSystemBLL.Entities;
 using KnowledgeTestingSystemBLL.Interfaces;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -11,23 +12,34 @@ namespace KnowledgeTestingSystem.Controllers
 {
     [ApiController]
     [Route("[controller]")]
+    [Authorize(Roles = "user, admin")]
     public class UserController : Controller
     {
         private readonly IUserService _userService;
+        private readonly IRoleService _roleService;
+        private readonly IAccountService _accountService;
+        private readonly IMapper _mapper;
 
-        public UserController(IUserService userService)
+        public UserController(IUserService userService, IRoleService roleService, IAccountService accountService)
         {
             _userService = userService;
+            _roleService = roleService;
+            _accountService = accountService;
+            var config = new MapperConfiguration(cfg =>
+            {
+                cfg.CreateMap<UserDTO, ApplicationUser>().ReverseMap();
+            });
+            _mapper = new Mapper(config);
         }
 
         [HttpGet("getUsers")]
+        [AllowAnonymous]
         public async Task<IActionResult> GetUsers()
         {
             return Ok(await _userService.GetAllAsync());
         }
 
         [HttpGet("me")]
-        [Authorize(Roles = "user")]
         public async Task<IActionResult> GetMeAsync()
         {
             string email = User.FindFirst(ClaimTypes.Name)?.Value;
@@ -41,23 +53,44 @@ namespace KnowledgeTestingSystem.Controllers
         [HttpDelete("deleteUser")]
         public async Task<IActionResult> DeleteUser(int id)
         {
-            var isDeleted = await _userService.DeleteAsync(id);
+            string email = User.FindFirst(ClaimTypes.Name)?.Value;
+            var user = await _userService.GetAsync(x => x.Email == email);
+            bool isAdmin = false;
+            foreach(var item in await _roleService.GetRoles(email))
+            {
+                if(item == "admin")
+                    isAdmin = true;
+            }
+            if (user.ElementAt(0).Id == id || isAdmin)
+            {
+                var isDeleted = await _userService.DeleteAsync(id);
 
-            if (!isDeleted)
-                throw new ArgumentException("You passed invalid user, it is not deleted");
-
-            return Ok();
+                if (!isDeleted)
+                    return BadRequest("You passed invalid user, it is not deleted");
+                return Ok();
+            }
+            else
+                return Forbid();
         }
 
         [HttpPut("editUser")]
         public async Task<IActionResult> EditUser(UserCompleteInformation newUser)
         {
             if (newUser == null)
-                throw new ArgumentNullException(nameof(newUser));
+                return BadRequest("You passed no user");
 
-            await _userService.EditCompleteAsync(newUser);
+            string email = User.FindFirst(ClaimTypes.Name)?.Value;
+            var user = await _userService.GetAsync(x => x.Email == email);
 
-            return NoContent();
+            if (user.ElementAt(0).Id == newUser.User.Id)
+            {
+                await _userService.EditCompleteAsync(newUser);
+
+                return NoContent();
+            }
+            else
+                return Forbid("You can't edit other user info");
         }
+
     }
 }
